chore: rename reveal to focal reveal,  reveal move to focal reveal move, simple reveal to reveal
