+++ packages/yubaba/src/__docs__/1-introduction/docs.mdx
@@ -19,7 +19,7 @@ easy to use animations with powerful orchestration for React.js 🧙✨
 It's all about **animation** 🧙✨ - it can help with:
 
 - [Moving an element](/move) from one position to another
-- [Revealing elements](/reveal-move) like you see above
+- [Revealing elements](/focal-reveal-move) like you see above
 - [Supporting animations](/supporting-animations) by obstructing elements in view; and
 - [Hiding children elements](/advanced-usage#delay-showing-content-until-all-animations-have-finished) until animations have completed to trick users 🤫
 - [Orchestrating](/advanced-usage#wait-for-the-previous-animation-to-finish-before-starting-the-next) when animations should happen and [in what order](/advanced-usage#controlling-in-what-order-animations-should-execute)
