+++ packages/yubaba/src/__docs__/3-advanced-usage/docs.mdx
@@ -29,7 +29,7 @@ import * as Styled from './styled';
 import { WrappedBaba as Baba } from '../../Baba';
 import { WrappedBabaManager as BabaManager } from '../../BabaManager';
 import CircleExpand from '../../animations/CircleExpand';
-import RevealMove from '../../animations/RevealMove';
+import FocalRevealMove from '../../animations/FocalRevealMove';
 import ConcealMove from '../../animations/ConcealMove';
 import FocalTarget from '../../FocalTarget';
 import Move from '../../animations/Move';
