+++ packages/yubaba/src/__docs__/focal-animations/docs.mdx
@@ -9,5 +9,5 @@ menu: Focal animations
 The bread and butter of `yubaba` 🍞.
 
 Focal animations are all about a single focal element that we want to transition from one state to another.
-While multiple elements _may_ be utilised in a focal animation (e.g. how [RevealMove](/reveal-move) works),
+While multiple elements _may_ be utilised in a focal animation (e.g. how [FocalRevealMove](/focal-reveal-move) works),
 the animations are always interested in a single focal element.
