+++ packages/yubaba/src/animations/FocalRevealMove/__docs__/images/halcyda.png
