+++ packages/yubaba/src/animations/FocalRevealMove/__docs__/styled.tsx
