+++ packages/yubaba/src/animations/Move/index.tsx
@@ -29,7 +29,7 @@ export interface MoveProps extends CollectorChildrenProps {
 
   /**
    * Will use <FocalTarget /> size and location for destination transform calculation.
-   * Internally this is used for the <RevealMove /> animation.
+   * Internally this is used for the <FocalRevealMove /> animation.
    */
   useFocalTarget: boolean;
