+++ packages/yubaba/src/animations/Reveal/index.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { css } from 'emotion';
 import Collector, {
   CollectorChildrenProps,
   AnimationCallback,
@@ -12,14 +11,11 @@ import { Duration } from '../types';
 
 export interface RevealProps extends CollectorChildrenProps {
   /**
+   * Takes either "dynamic" or a number in ms.
    * How long the animation should take over {duration}ms.
+   * Defaults to "dynamic".
    */
   duration: Duration;
-
-  /**
-   * zIndex to be applied to the moving element.
-   */
-  zIndex?: number;
 
   /**
    * Timing function to be used in the transition.
@@ -27,81 +23,42 @@ export interface RevealProps extends CollectorChildrenProps {
   timingFunction: string;
 
   /**
-   * Set to false to disable transforming the children to the X position of the focal element.
-   * Defaults to true.
+   * This will increase (or decrease) the size of the clip-path box.
+   * Use with caution.
+   * [top, right, bottom, left]
    */
-  childrenTransformX: boolean;
-
-  /**
-   * Set to false to disable transforming the children to the Y position of the focal element.
-   * Defaults to `true`.
-   */
-  childrenTransformY: boolean;
-
-  /**
-   * Will transition using clip-path over height and width.
-   * This results in a more resilient transition since page flow isn't affected at the cost of browser support.
-   * See: https://caniuse.com/#feat=css-clip-path
-   */
-  useClipPath?: boolean;
+  offset: [number, number, number, number];
 }
 
 export default class Reveal extends React.Component<RevealProps> {
   static defaultProps = {
     duration: 'dynamic',
     timingFunction: standard(),
-    childrenTransformX: true,
-    childrenTransformY: true,
+    offset: [0, 0, 0, 0],
   };
 
   beforeAnimate: AnimationCallback = (data, onFinish, setChildProps) => {
-    if (!data.destination.focalTargetElementBoundingBox) {
-      throw new Error(`yubaba
-<FocalTarget /> was not found, if you haven't defined one make sure to add one as a descendant of your target Baba.`);
-    }
+    const [topOffset, rightOffset, bottomOffset, leftOffset] = this.props.offset;
 
-    const { childrenTransformX, childrenTransformY, useClipPath } = this.props;
-
-    const offsetChildrenX = childrenTransformX
-      ? data.destination.focalTargetElementBoundingBox.location.left -
-        data.destination.elementBoundingBox.location.left
-      : 0;
-    const offsetChildrenY = childrenTransformY
-      ? data.destination.focalTargetElementBoundingBox.location.top -
-        data.destination.elementBoundingBox.location.top
-      : 0;
-
-    const revealStyles = useClipPath
-      ? {
-          clipPath: `inset(0 ${data.destination.elementBoundingBox.size.width -
-            data.destination.focalTargetElementBoundingBox.size.width}px ${data.destination
-            .elementBoundingBox.size.height -
-            data.destination.focalTargetElementBoundingBox.size.height}px 0)`,
-        }
-      : {
-          height: data.destination.focalTargetElementBoundingBox.size.height,
-          width: data.destination.focalTargetElementBoundingBox.size.width,
-        };
+    const right =
+      data.destination.elementBoundingBox.size.width -
+      data.origin.elementBoundingBox.size.width +
+      rightOffset;
+    const bottom =
+      data.destination.elementBoundingBox.size.height -
+      data.origin.elementBoundingBox.size.height +
+      bottomOffset;
 
     setChildProps({
       style: prevStyles =>
-        data.destination.focalTargetElementBoundingBox
+        data.origin.elementBoundingBox
           ? {
               ...prevStyles,
-              ...revealStyles,
+              clipPath: `inset(${topOffset}px ${right}px ${bottom}px ${leftOffset}px)`,
               opacity: 1,
               visibility: 'visible',
-              willChange: combine('height, width, clip-path')(prevStyles.willChange),
-              overflow: 'hidden',
+              willChange: combine('clip-path')(prevStyles.willChange),
             }
-          : undefined,
-      className: () =>
-        data.destination.focalTargetElementBoundingBox
-          ? css({
-              '> *': {
-                transform: `translate3d(-${offsetChildrenX}px, -${offsetChildrenY}px, 0)`,
-              },
-            })
           : undefined,
     });
 
@@ -109,36 +66,20 @@ export default class Reveal extends React.Component<RevealProps> {
   };
 
   animate: AnimationCallback = (data, onFinish, setChildProps) => {
-    const { timingFunction, duration, useClipPath } = this.props;
+    const { timingFunction, duration } = this.props;
     const calculatedDuration =
       duration === 'dynamic'
         ? dynamic(data.origin.elementBoundingBox, data.destination.elementBoundingBox)
         : duration;
 
-    const revealStyles = useClipPath
-      ? {
-          clipPath: 'inset(0 0 0 0)',
-        }
-      : {
-          height: data.destination.elementBoundingBox.size.height,
-          width: data.destination.elementBoundingBox.size.width,
-        };
-
     setChildProps({
       style: prevStyles => ({
         ...prevStyles,
-        ...revealStyles,
-        transition: combine(
-          `height ${calculatedDuration}ms ${timingFunction}, width ${calculatedDuration}ms ${timingFunction}, clip-path ${calculatedDuration}ms ${timingFunction}`
-        )(prevStyles.transition),
+        clipPath: 'inset(0px)',
+        transition: combine(`clip-path ${calculatedDuration}ms ${timingFunction}`)(
+          prevStyles.transition
+        ),
       }),
-      className: () =>
-        css({
-          '> *': {
-            transform: `translate3d(0, 0, 0)`,
-            transition: `transform ${calculatedDuration}ms ${timingFunction}`,
-          },
-        }),
     });
 
     setTimeout(() => onFinish(), calculatedDuration);
