--- packages/yubaba/src/animations/SimpleReveal/__docz__/docs.mdx
@@ -1,30 +1,0 @@
----
-name: SimpleReveal
-route: /simple-reveal
-menu: Focal animations
----
-
-import { Playground, PropsTable } from 'docz';
-import * as Common from 'yubaba-common';
-import SimpleReveal from '../index';
-
-# SimpleReveal
-
-Will reveal the destination element from the origin size to the destination size using `clip-path` only which will not work for IE11.
-
-## Usage
-
-```js
-import { SimpleReveal } from 'yubaba';
-```
-
-## Props
-
-<PropsTable of={SimpleReveal} />
-
-## Caveats
-
-Be careful of collapsing margins when utilising this animation,
-they will make the destination element jump around,
-probably.
-If you're seeing some odd behavior - maybe try a flex container instead.
